made the rabbitmq functional inside the message module.
